requirements.txt
web3>=6.20
py-solc-x>=2.0
cryptography>=42.0.0
